remove ``on_artwork_stream_start`` and ``on_artwork_stream_end`` from the artwork listener API. These are both unnecessary. The server guarantees it will send us the format specified in the hello message. The other callback methods handle all the details necessary for a client to implement artwork following the Sendspin spec.
