+++ src/client.cpp
@@ -484,7 +484,7 @@ void SendspinClient::process_json_message(SendspinConnection* conn, const std::s
 
 #ifdef SENDSPIN_ENABLE_ARTWORK
             if (this->artwork_ && stream_msg.artwork.has_value()) {
-                this->artwork_->impl_->handle_stream_start(stream_msg.artwork.value());
+                this->artwork_->impl_->handle_stream_start();
             }
 #endif
